:sparkles: [feat] 검색어 입력시 화면 전환 기능 개발, [refactor] 검색화면 코드 리펙토링 (HH-365)
